Conflicts:
	build_windows/Berkeley_DB.sln
	build_windows/Berkeley_DB_vs2010.sln

+++ examples/c/csv/DbRecord.c
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 2005, 2011 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 2005, 2012 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -369,7 +369,7 @@ DbRecord_search_field(DbField *f, char *value, OPERATOR op)
 		case STRING:
 			valuep = value;
 			cmp = field_cmp_string;
-			key.size = strlen(value);
+			key.size = (u_int32_t)strlen(value);
 			break;
 		case UNSIGNED_LONG:
 			if (strtoul_err(value, &value_ulong) != 0)
